+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `summary` property in Article objects, using a configurable post property with rendered excerpt fallback.
 - Note object generation for short, untitled posts, with configurable character limit.
 - `url` property in Actor objects, to link to the site home page.
+- `page.activitypub_url` for the site index, to link the home page to its ActivityPub Actor object.
 
 ## [0.7.6] - 2026-09-04